feature Construction ONLY, has been modularised
