+++ requirements.txt
@@ -5,3 +5,4 @@ tqdm>=4.66.1
 spacy>=3.7.2
 emojis>=0.7.0
 feel-it>=1.0.5
+pyarrow
